rename maybe-remove-ad-slot -> manage-ad-free-cookie-on-consent-change
